resume_analysis_project/requirements.txt
streamlit
PyPDF2
python-docx
pytesseract
Pillow
scikit-learn
pandas
joblib